+++ src/test/converter2/issues/gh1973/a.ts
@@ -1,0 +1,9 @@
+/**
+ * A module
+ * @module
+ */
+
+/**
+ * Comment for a
+ */
+export const a = "a";