+++ src/test/converter2/issues/gh1973/b.ts
@@ -1,0 +1,9 @@
+/**
+ * B module
+ * @module
+ */
+
+/**
+ * Comment for b
+ */
+export const b = "b";