+++ src/test/converter2/issues/gh1973/index.ts
@@ -1,0 +1,7 @@
+/**
+ * A override
+ */
+export * as A from "./a.js";
+
+// No override
+export * as B from "./b.js";